feat: add tilde lines, fix uninitialized state, expose buffer access
- Added tilde (~) lines as editor row references, dynamically adjusted to screen height.
- Removed uninitialized state usage.
- Added buffer return/access function.

+++ src/core/editor.hpp
@@ -1,12 +1,14 @@
 #ifndef EDITOR_H
 #define EDITOR_H
 
+#include <windows.h>
+
 struct EditorState{
 	int row;
 	int col;
 	int cursor_line;
 	int cursor_col;
-	int originalColor;
+	WORD originalColor;
 };
 
 class Editor{
